+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to `Weasley` will be documented in this file.
 
 ## [0.2.0](https://github.com/rougin/weasley/compare/v0.1.2...v0.2.0) - Unreleased
 
+### Added
+- `JSON_PARTIAL_OUTPUT_ON_ERROR` on `json_encode` in `BaseController::json`
+
 ### Changed
 - `BaseController::toJson` to `BaseController::json`
 - DocBlocks documentation on `RestfulController`